Add `FlexVec`, rename `ParticleSpawner` to `World`, fix some docstrings
`FlexVec<T>` is essentially a wrapper of `NvFlexBuffer`, similar to the
`NvFlexVector` struct in `NvFlexExt`. It has some methods that are
similar to the ones in `Vec`. `ParticleSpawner` was renamed to `World`
because it will store all `FlexVec`s required for the simulation,
making it easier to add particles/colliders/etc to it.